Rework structure for CV with more detail
Move all content into dedicated files and replace main.tex with files for
cv and cv-long. These take different mixtures of the content files to create
the longer and shorter versions of the CV document.

The longer version includes a personal statement and extra projects.

Change formatting rules for the projects list to so each is separated
by a set ammount rather than the list expanding to fill all available
space.